pyproject.toml
[project]
name = "thechessbuddy"
version = "0.1.0"
description = "Chess buddy is an AI agent that can help you get better at chess."
requires-python = ">=3.13"
dependencies = [
    "modal>=0.74.29",
    "openai>=1.76.0",
    "smolagents[mcp]>=1.14.0",
    "requests>=2.31.0",
    "pytest>=8.0.0",
    "chess-com>=3.11.1",
    "python-chess>=1.999",
    "pandas>=2.2.3",
    "numpy>=2.2.5",
    "streamlit>=1.44.1",
    "logfire[fastapi,requests,system-metrics,starlette]>=3.14.1,<4.0.0",
    "pydantic-ai>=0.1.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["komodo"]

[tool.uv.workspace]
members = [
    "chessbuddy",
]
